chore: alter warning method and reorganize code
